added  : base geometry instancing (just the draw logic) changed: gl callback option from UInt32 to UInt64        : exclude by name search trav mask from default rendering masks

+++ Source/System/NodeCores/Drawables/Geometry/PropertiesBase/OSGGeoProperty.h
@@ -187,7 +187,7 @@ class OSG_SYSTEM_DLLMAPPING GeoProperty : public GeoPropertyBase
            UInt32 handleGL       (DrawEnv                 *pEnv, 
                                   UInt32                   id, 
                                   Window::GLObjectStatusE  mode,
-                                  UInt32                   uiOptions);
+                                  UInt64                   uiOptions);
     static void   handleDestroyGL(DrawEnv                 *pEnv, 
                                   UInt32                   id, 
                                   Window::GLObjectStatusE  mode     );
