+++ Source/System/State/OpenGL/OSGCubeTextureObjChunk.h
@@ -177,7 +177,7 @@ class OSG_STATE_DLLMAPPING CubeTextureObjChunk : public CubeTextureObjChunkBase
            UInt32 handleGL       (DrawEnv                 *pEnv, 
                                   UInt32                   id, 
                                   Window::GLObjectStatusE  mode,
-                                  UInt32                   uiOptions);
+                                  UInt64                   uiOptions);
     static void   handleDestroyGL(DrawEnv                 *pEnv, 
                                   UInt32                   id, 
                                   Window::GLObjectStatusE  mode);
